修改double函数为float这样效率更高，因为学生成绩不可能那么大大保留三位数就足够来

+++ C语言之函数/函数/函数/get_average.c
@@ -9,14 +9,14 @@
 #include "get_average.h"
 int accelerateaverage(){
     int count=0;
-    double average_score=0.0,total_score=0.0,score=0.0;
+    float average_score=0.0,total_score=0.0,score=0.0;
     printf("请输入学生的成绩:");
-    scanf("%lf",&score);
+    scanf("%f",&score);
     while (score>=0) {
         count++;
         total_score+=score;
         printf("请输入学生的成绩:");
-        scanf("%lf",&score);
+        scanf("%f",&score);
     }
     average_score=total_score/count;
     printf("学生总人数为:%d\n学生总分数:%.3f\n平均分为:%.3f\n",count,total_score,average_score);
